+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+# Version 2.0.0
+
+*Contains breaking changes*
+
++ No longer a single file, it now namespaced under `OhmyBrew\ShopifyAPI`
++ GraphQL is now introduced under `OhMyBrew\ShopifyAPI\GraphAPI`
++ REST is moved to `OhMyBrew\ShopifyAPI\RestAPI`
+
 # Version 1.0.1
 
 + Fixed issue #3 by @ncpope, for newly created stores (after December 2017) not accepting GET body